requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
python-dotenv
pydantic-settings
rich
asgi-correlation-id
python-json-logger
logtail-python
python-jose
python-multipart
passlib[bcrypt]
boto3[crt]
aiofiles
b2sdk
